implement scanning for vst plugins on windows
the library now reads a complete list of VST addons when starting
initially.

this approach is blocking and quite slow, but should work
best, when it comes to loading VST shell addons and having a
complete list of all addons.

this also has the benefit of only showing addons that can be loaded,
not like the original implementation that just refuses to load a
addon, if it is unable to open it during the runtime.

so the next step would be cleaning stuff properly up and binding
the collected data to the vst source.